Fix an issue with PSD estimation in notebook
Thanks to Alex Neilsen for pointing out to me that
I had been using GW170104 PSDs for whitening the data.
For Hanford this basically makes no difference as the PSD
shape is largely unchanged. For Livingston this improves
the ability to whiten the data and further removes correlations.
Plots are remade with this fix in place.