Changes in repspone to Sebastian von Hausegger
In response to some comments, and issues, pointed out
by Sebastian von Hausegger. The main difference in
terms of results from this patch is that I now am
able to produce the behaviour in Figure 7 when using the
residual data for Figure 1 of the PRL. I don't believe
this peak is significant, but I agree it is worth
some effort to try to understand. Especially as it is
not reproduced in the whitened data.